image-remove-alpha-channel/requirements.txt
requests==2.22.0
Pillow==6.1.0
datetime==4.3